Dampens overshoot, adds straight driving
This commit vastly dampens the overshoot the robot has when turning by
setting the speed as a power of the angle of the turntarget. Also it
adds a compile option (which is kind off hacked into a feature where
you can switch the kind driving with the joystick 8 button) to disable
compass compensation, so that the front of the robot always stays the
front for the driving.